+++ src/test/java/com/appium/platforms/ios/IOSDeviceActionsTest.java
@@ -1,5 +1,6 @@
 package com.appium.platforms.ios;
 
+import com.appium.models.BatteryInfoModel;
 import io.appium.java_client.AppiumDriver;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
@@ -9,8 +10,7 @@ import org.mockito.MockitoAnnotations;
 
 import java.util.Map;
 
-import static org.junit.jupiter.api.Assertions.assertFalse;
-import static org.junit.jupiter.api.Assertions.assertTrue;
+import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 class IOSDeviceActionsTest {
@@ -82,5 +82,16 @@ class IOSDeviceActionsTest {
         when(driver.executeScript("mobile: isLocked")).thenReturn(false);
         assertFalse(iosDeviceActions.isLocked(driver));
         verify(driver, times(1)).executeScript("mobile: isLocked");
+    }
+
+    @DisplayName("Should return BatteryInfoModel when batteryInfo is called")
+    @Test
+    void testBatteryInfo() {
+        when(driver.executeScript("mobile: batteryInfo")).thenReturn(Map.of("level", 1, "state", 2));
+        final BatteryInfoModel batteryInfo = iosDeviceActions.batteryInfo(driver);
+
+        assertEquals(1, batteryInfo.getLevel());
+        assertEquals(2, batteryInfo.getState());
+        verify(driver, times(1)).executeScript("mobile: batteryInfo");
     }
 }