+++ CHANGELOG.md
@@ -8,7 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0] - 2024-07
 
 ### **Added**
-  - deepLink actions
+  - Session actions
+    - deepLink
+  - Device actions
+    - batteryInfo
 
 ## [1.0.0] - 2024-07-09